+++ R/coursekata_attach.R
@@ -1,5 +1,6 @@
 #' Attach the CourseKata course packages
 #'
+#' @param do_not_ask Prevent asking the user to install missing packages (they are skipped).
 #' @param quietly Whether to suppress messages.
 #'
 #' @return A named logical vector indicating which packages were attached.
@@ -7,9 +8,15 @@
 #' @export
 #' @examples
 #' coursekata_attach()
-coursekata_attach <- function(quietly = FALSE) {
-  to_attach <- coursekata_detached()
-  invisible(pkg_require(to_attach, quietly = quietly))
+coursekata_attach <- function(do_not_ask = FALSE, quietly = FALSE) {
+  !do_not_ask && pkg_check_installed(coursekata_pkgs)
+  detached <- coursekata_detached()
+  installed <- coursekata_pkgs[pkg_is_installed(coursekata_pkgs)]
+  attached <- pkg_require(detached[detached %in% installed], quietly = quietly)
+
+  result <- rep_named(detached, FALSE)
+  result[names(attached)] <- TRUE
+  invisible(result)
 }
 
 
@@ -23,7 +30,9 @@ coursekata_attach <- function(quietly = FALSE) {
 #'
 #' @noRd
 coursekata_attach_message <- function(pkgs) {
-  if (length(pkgs) == 0) return(NULL)
+  if (length(pkgs) == 0) {
+    return(NULL)
+  }
 
   info <- coursekata_packages()
   version <- ifelse(is.na(info$version), "", info$version)
@@ -41,4 +50,14 @@ coursekata_attach_message <- function(pkgs) {
     to_cols(pkgs, 2),
     sep = "\n"
   )
+}
+
+
+#' List all currently NOT attached CourseKata course packages
+#'
+#' @return A character vector of the course packages that are not attached.
+#'
+#' @noRd
+coursekata_detached <- function() {
+  coursekata_pkgs[!pkg_is_attached(coursekata_pkgs)]
 }