+++ R/zzz.R
@@ -1,5 +1,9 @@
 .onAttach <- function(...) {
-  attached <- coursekata_attach(quietly = getOption("coursekata.quiet", FALSE) || quickstart())
+  attached <- coursekata_attach(
+    do_not_ask = !interactive() || interactive() && quickstart(),
+    quietly = getOption("coursekata.quiet", FALSE) || quickstart()
+  )
+
   coursekata_load_theme()
   if (!quickstart()) {
     rlang::inform(coursekata_attach_message(attached), class = "packageStartupMessage")