+++ tests/testthat/test-utils-pkg.R
@@ -1,21 +1,20 @@
 test_that("it determines whether a package is attached or not", {
-  pkgs <- "fivethirtyeight" # use this package because it is not imported
+  # use a package that can be unloaded and reloaded (so, can't be imported by coursekata)
+  pkgs <- "fivethirtyeight"
+
+  # make sure the package can be added
+  if (!require(pkgs, character.only = TRUE, quietly = TRUE)) {
+    fail(paste("Package not available:", pkgs))
+  }
 
   purrr::walk(pkgs, detacher)
-  expect_vector(pkg_is_attached(pkgs), logical(), length(pkgs))
-  expect_true(all(!pkg_is_attached(pkgs)))
+  attached <- pkg_is_attached(pkgs)
+  expect_vector(attached, logical(), length(pkgs))
+  expect_true(all(!attached))
 
   # reattach for other tests
   purrr::walk(pkgs, attacher)
   expect_true(all(pkg_is_attached(pkgs)))
-})
-
-
-test_that("it retrieves the library location for currently installed packages, or NA", {
-  pkgs <- c("supernova", "lsr", "does_not_exist")
-  locations <- pkg_library_location(pkgs)
-  expect_identical(unname(locations[3]), NA_character_)
-  expect_true(all(dir.exists(locations[1:2])))
 })
 
 
@@ -39,20 +38,6 @@ test_that("requiring a package can be quiet", {
   expect_message(pkg_require(pkgs, quietly = TRUE), NA)
 })
 
-test_that("requiring a missing package calmly asks if you want to install it", {
-  skip("Make sure to run this when testing locally.")
-
-  menu_mock <- mockery::mock(FALSE)
-  mockr::with_mock(
-    .env = as.environment("package:coursekata"),
-    ask_to_install = menu_mock,
-    {
-      expect_warning(pkg_require("does not exist"), NA)
-      expect_length(mockery::mock_calls(menu_mock), 1)
-    }
-  )
-})
-
 
 test_that("installing fivethirtyeightdata works as intended", {
   skip("Make sure to run this when testing locally.")
@@ -64,6 +49,6 @@ test_that("installing fivethirtyeightdata works as intended", {
     error = function(e) invisible(NULL)
   )
 
-  pkg_install("fivethirtyeightdata")
+  pkg_check_installed("fivethirtyeightdata")
   expect_true(require("fivethirtyeightdata"))
 })